nstemplate test coverage

+++ yamlns/nstemplate_test.py
@@ -3,6 +3,9 @@
 import unittest
 import yamlns
 import sys
+import io
+import os
+from . import nstemplate
 
 class NSTemplate_test(unittest.TestCase) :
 
@@ -14,8 +17,16 @@ class NSTemplate_test(unittest.TestCase) :
 
 	def tearDown(self) :
 		for f in self._toRemove :
-			os.remove(f)
+			if os.path.exists(f):
+				os.remove(f)
 
+	def write(self, filename, content):
+		with io.open(filename, 'w', encoding='utf8') as f:
+			f.write(content)
+
+	def read(self, filename):
+		with io.open(filename, encoding='utf8') as f:
+			return f.read()
 
 	def test_collectVars_withNoVar(self) :
 		content = "booo"
@@ -122,6 +133,54 @@ class NSTemplate_test(unittest.TestCase) :
 			'    far: \'\'\n'
 			'    nice: \'\'\n'
 			)
+
+	def test_templateVarsAsYaml(self):
+		yaml = yamlns.namespace.fromTemplateVars("""\
+			{var1.foo} {var1.lala} {var2}
+		""")
+		self.assertEqual(yaml.dump(), 
+			'var1:\n'
+			'  foo: \'\'\n'
+			'  lala: \'\'\n'
+			'var2: \'\'\n'
+		)
+
+	def test_extract(self):
+		filemd = 'deleteme.md'
+		fileyaml = 'deleteme.yaml'
+		self.toRemove(filemd)
+		self.toRemove(fileyaml)
+		self.write(filemd, u"{var1} {var2.foo}")
+
+		nstemplate.extract(filemd, fileyaml)
+
+		self.assertEqual(self.read(fileyaml),
+			u"var1: ''\n"
+			u"var2:\n"
+			u"  foo: ''\n"
+		)
+
+	def test_apply(self):
+		filemd = 'deleteme.md'
+		fileyaml = 'deleteme.yaml'
+		fileout = 'deleteme-output.md'
+		self.toRemove(filemd)
+		self.toRemove(fileyaml)
+		self.toRemove(fileout)
+
+		self.write(filemd, u"{var1} {var2.foo}")
+		self.write(fileyaml,
+			u"var1: 'value1'\n"
+			u"var2:\n"
+			u"  foo: 666\n"
+		)
+
+		nstemplate.apply(fileyaml, filemd, fileout)
+
+		self.assertEqual(self.read(fileout),
+			u"value1 666")
+
+
